[AIP-21] Move contrib.operators.vertica_to_hive providers.apache.hive.operators.vertica_to_hive
